Refrain from selecting a prim/item in the PrimView when interacting with the Vis or DrawMode columns.
We cannot override the selection behavior of a QtTreeWidget directly -
you either select entire rows, or entire columns, or individual cells,
but there is no way to state that only certain cells in a row should
trigger selection.

So instead we specialize the Widget's QItemSelectionModel to allow us
to configure it to ignore selection requests, and do so between
mousePress and mouseRelease events when the mousePress hit one of the
columns we want to be "unselectable".

Since Vis'ing no longer causes selection, it's incorrect for the
appController to update the PropertyView with the focus prim, and in
fact it was already unnecessary since the whole UI is (as of recently)
responsive to stage edits.  So we remove the forecful update.

Finally, make it so that if we click the Vis column of a *selected* prim,
we will toggle visibility for *all* selected prims.

(Internal change: 1932843)