Detect Python to link USD against in build_usd.py
build_usd.py will now query the version of Python found
in the user's PATH to determine the include directory and
shared library to use when building USD.

This is intended to help cases where a user has multiple
versions of Python installed on their system. In such
cases, CMake can get confused and wind up linking USD
against a different Python installation than the one
that would normally be used, which can lead to crashes.

In particular, this has come up on macOS when a user
has a homebrew-based Python installation that conflicts
with the system-provided Python install. In this case,
CMake would link USD against the system Python even
though the homebrew Python is the one that the user
actually invokes. This would lead to a crash when
trying to use any of the USD Python bindings.

Note that this behavior is only invoked on Linux and
macOS. The method for determining the Python include
directory and library doesn't work on Windows, and
it seems much more difficult/unlikely to have multiple
installations of the same Python version on that
platform.

Thanks to Rob Pieke for the idea behind this change!

Fixes #19
Fixes #699

(Internal change: 1932880)