Check Visual Studio version via environment variable
build_usd.py had previously tried to parse the output of
cl.exe to retrieve the Visual Studio version number.
However, the version string in this output appears to
be locale-specific, so this failed on non-English
platforms.

To avoid this, we now parse the VisualStudioVersion
env var that is set as part of the developer command
prompt.

Fixes #613
Fixes #697

(Internal change: 1933169)