Skip Python detection on Linux in build_usd.py
The Python detection code appears to give incorrect
answers on Ubuntu: the code thinks the Python library
lives in /usr/lib, but it actually lives in a different
directory according to the multiarch specification. The
same issue would presumably occur on any distribution
implementing that spec, like Debian.

Since the problem this code was intended to solve
primarily occurs on macOS, this fix simply skips over
the Python detection on Linux.

See #19
See #699

(Internal change: 1933560)