- Make sure alpha_to_coverage are off during composite and color-correction.
- Make sure depth test is ALWAYS during color-correction and ImagingGL composite back to client buffers.
  Disabling depth-test also disabled depth writes.
  We need to copy the depth buffer to client buffer, but do not want any depth testing to cause some pixels not to be copied.

Alpha_to_coverage 'could' potentiallty be interesting to use for composite in the future, but then we
need to explicitely let the user of the composite enable this option.
Right now any pixel failing the depth-test is ignored during composite, which seems to be more intuitive then using alpha_to_coverage.

(Internal change: 1933790)