Re-structured how familyName and familyType are handled.
Alembic has no concept of familyName, so "materialBind" was used as the default family name here.
In order to modify both the GeomSubset (to add the indices, etc) and the parent prim (to add the familyType), some modifications were made to the _PrimReaderContext and _PrimWriterContext to allow access to parent primitives.