[Hd] Split hydra's "color" primvar into a float3 "displayColor" and a float "displayOpacity".
This is done for better compatibility with raytracers (which consume the values separately), and better compatibility with the way the data is provided in both usd.  Typical scenes have constant-rate opacity per mesh, which means splatting it across the color values is a waste of space, and combining the two primvars makes it next to impossible to properly implement inherited color and opacity, including taking instance-rate values into account.

(Internal change: 1937253)