[UsdMaya] Remove special-case generation of single-axis transform ops (rotateX, rotateY, rotateZ)
The single-axis rotate ops save a bit of space, but ultimately cause prims to be
incompatible with UsdGeomXformCommonAPI. This is often a gotcha for downstream
consumers that expect prims to be common API-compatible. The single-axis rotates
are now output as just the three-axis rotates with only a single nonzero-component.

This change updates testUsdExportCamera to deal with the different xformOps,
and also checks that the prims are common API-compatible.

(Internal change: 1937260)