[Hd/Hdx] Add Prepare Phase for Tasks.
HdTask's Sync function has been doing double duty.  It works like other sync functions, in that it pulls task state from the scene delegate.  However, in some task sync also does work to set up the task context and resources needed for a specific frame.  As a result, there is an assumption that task sync runs on every call to "Execute".

This function overloading, also leads to a circular dependency in sync management.  Tasks have links to Sprims (such as camera and lights), which need resolving before the task, but the task manages the collection that tells us what to sync.

To break this dependency loop and clean up this double duty, a new "Phase" in introduced for tasks.  The "Prepare" phase is guaranteed to be executed for every task in the current task list (Sync may not be called in the future if the task isn't dirty) and runs after Sync phase (so all prims including Sprims have been Synced), but before the Commit Phase (so tasks Prepare phase can add resources).

Changes:
- Added Prepare() virtual method to HdTask.
- Provided empty Implementation of Prepare method in all tasks.

Notes:
Migrating tasks to using Prepare has varying complexities, so will be done as separate changes.

(Internal change: 1937983)