Removing the need for polymorphic python wrapping in Pcp specifically for PcpPayloadDecorator.
Added new python wrapper to Pcp for a python only test scope called
_TestPayloadDecorator. This scope has one function that creates a special
PcpCache for testing that a special PcpPayloadDecorator just for testing in
testPcpPayloadDecorator.

Deleted the python wrappings for PcpPayloadContext and PcpPayloadDecorator and
removed the payloadDecorator argument from PcpCache's wrapped init. There was no
way to construct or get a PcpPayloadContext in python and we didn't need payload
decorators exposed in python other than for than the test.

Updated the testPcpPayloadDecorator to use _TestPayloadDecorator to create its
test cache and decorator as well as fixing the test case in test_Changes that
actually giving false positives.

(Internal change: 1938911)