[hdx/hdSt] Pass render index to GetDrawTargets()
Under the previous design the Sync phase was run every frame.  Under the new 3-phase scheme, the Prepare phase fulfills this role and in the future, Sync may only be run if the task is dirty. Under prepare, only the render index is passed as the task shouldn't be communicating with the delegate at this phase (delegate communication should only take place in sync).

Changes:
- HdStDrawTarget's GetDrawTargets() static method now takes a pointer to the render index instead of the scene delegate.

(Internal change: 1939338)