[Hd] Fix for Instancer Creation Failure.
While modifying a test to use the Null Render Delegate, the test was segfaulting while removing instancers from the render index.
The reasons for this was that the Null Render Delegate was failing to create the instancer (which it is allowed to do), but the render index was were not catching the error and inserting a null entry, which was dereferenced in the remove code.

It is currently the policy, that if an error occurs creating a prim, that prim does not get inserted into the render index.

So this change, applies this policy to Instancer prims to.

Changes:
- Catch a nullptr return from HdRenderIndex::CreateInstancer() and early return, skipping render index insertion.

(Internal change: 1940709)