[Hd] Upgrade Null Render Delegate to pull scene delegate data for rprims.
While modifying some test to use the Null Render Delegate, the tests were failing, because they were depending on perf counters that were not being updated by the null delegate.

These tests are white box scene delegate tests, expecting the render delegate to do Sync processing.  This is currently working fine as they use the stream delegate, but these tests
are not supposed to have a OpenGL dependency.

So this change, modifies the Null Rprim, so that acts like a Rprim, by checking dirty bits and in pulling data from the scene delegate during sync.  However, the resulting data isn't used.

In addition, the Null Render Delegate, now returns a valid instancer (using the basic implementation), as this was causing additional test failures.

Changes:
- Hd_UnitTestNullRenderDelegate::CreateInstancer() now returns a new HdInstancer.
- Hd_NullRprim now know's it's type (as the Topology scene delegate API is type dependent)
- Hd_NullRprim Sync, now checks dirty bits and if dirty calls the apropreate scene delegate getter.  When available the ChangeTracker Is*Dirty() functions are used, so that the "cache" perf counters are updated.
- For primvar dirtiness,  the null delegates obtains the list of all the primvars and invokes the getter for each.

Notes:

Subdiv tags are not pulled as they are only valid for refined prims.

(Internal change: 1940711)