[UsdImaging] Fix for hidden GL dependencies in UsdImaging tests.
The tests in the UsdImaging were depending on the stream render delegate to execute.  Due to the way the backend ran, it introduced an indirect GL dependency and changes to Glf were being exposed to these tests.

Rather than introduce more ties to GL, this change switches these tests over to use the Null Render Delegate.  Removing the dependency for the HdSt library and Glf and GL.

Primarally, UsdImaging's unitTestHelper now uses the null delegate.  This resulted in quite a bit of dead code around the camera and render pass state.

Changes:
- Removed pulling in of HdSt headers
- Renamed UsdImaging_DrawTask to UsdImaging_TestTask as it doesn't draw anything.
- UsdImaging_TestTask Prepare and Execute functions are now no-ops.
- Removed the render pass state.
- Removed SetCamera() as that was only needed for render pass state.
- Removed unused functions SetRefineLevelFallback() and GetRenderPass().

(Internal change: 1941410)