[Hd] Fix for invalidated iterator accesses in SortedId.
During HdSortedId's id removal code, it was using the iterator of the id's position in the list of all iterators, to determine how id's in the list remain sorted.

However, as this code removes the element in the list, if id happened to be the last element, the iterator is invalid.

As the operation performed is a distance operation (-) between the two iterators, Linux/Gcc seems to be ok, but a debug build on Windows does throw an iterator error.

This change fixes the code so it no longer access the invalid iterator.

Changes:
- Moved the calculation of _sortedCount, inside the check for the last element.
- The computation remains the same if the id is not the last element.
- If it is the last element _sortedCount is clamped to the new size of the list.

Fixes #731

(Internal change: 1941739)