Added initRender() called by both renderPick() and render()
renderPick() was written to assume render() had already been called on the
same data, but a crash stack dump seems to show that it is possible for it to
be called first. This puts the necessary initialization into a method shared
by both of them.

Unfortunately we have not been able to recreate the crash in the unpatched
version, so there is no proof this fixes anything.

Signed-off-by: Bill Spitzak <bill.spitzak@dreamworks.com>