[vt] Change Python repr for legacy shaped arrays in preparation for removing the shape data completely
This makes the __repr__ for shaped VtArrays explicitly non-eval()-able by printing it in a format like:
    <Vt.IntArray(4, (0, 1, 2, 3)) with shape (2, 2)>
instead of the previous format:
    Vt.IntArray((2, 2), (0, 1, 2, 3))
which was eval()-able, but would have errored in the constructor.

The new __repr__ makes uses the <angle bracket syntax> to indicate that it is non-eval() able.
(We will eventually remove the shaped repr once we strip out all remaining usage of the shaped
array feature completely.)

(Internal change: 1945928)