Introduces the concept of 'MaterialTag' into Hydra.
A MaterialTag can be used to filter HdRprimCollection based on material tags found on HdMaterials.
The goal is to be able to create different collections for opaque and translucent prims for SOME backends.

The rendering backend in the end has control over what material tags it wishes to support by choosing to pass the MaterialTags found on the HdMaterial to the DrawItem.

E.g.
The Stream backend wants DrawItems with 'translucent' tags to be split into a different collection so it can render translucent items in a second task.
While a path-tracer backend wants opaque and translucent prims to remain in the same collection.

This currently depends on the HdxRenderSetupTask pushing multiple renderTasks for known MaterialTags.
However, it that feels a bit hardcoded, it might be nice to query the backend somehow to request what RenderTasks should be pushed?

(Internal change: 1946300)