[UsdViewq] Fix for no GL context during UsdImagingGLEngine initialization.
UsdView's stage view class contains an instance of UsdImagingGLEngine it call's the _renderer.

The actual creation of the engine is deferred until the first call to _getRenderer().  Because it could be GL drawing or Picking
that can cause the render to be created.

However, in the GL drawing code path, the first call to _getRenderer() happens in the override of Qt's glDraw().
At this point, Qt has not created and set the context current.  Therefore, this call to _getRenderer() was attempting
to create a UsdImagingGLEngine without a valid gl context bound.  This was causing errors in UsdImagingGLEngine
initialization which needs to query the properties of the gl context.

The reason for the check in glDraw(), is to make sure no errors occur while creating the UsdImagingGLEngine and the
timer truly represents the time to create the first image.

Changes:
- Changes to check for the valid render from _getRenderer() to checking the variable directly, to avoid accidental creation.
- Moved the check until after the call to the base class draw (The paintGL function, which does have a valid context, now causes the creation).

Fixes #764

(Internal change: 1946472)