[Glf +] Fix for crash initializing GlfContextCaps.
Hydra deferred initialization of GlfContextCaps to the first time it was used.  This was typically ok for most users, as the task sync was the first to use the caps.
As task sync is single threaded, it runs on the thread HdEngine::Execute is called on, so it has the GL context bound.

With the changes to the 3 phase task system, task sync is no longer accessing the caps.  As a result, the first access to the caps has moved to Rprim sync for some workflows.
As Rprim sync is multi-threaded, it becomes a race to see which thread initializes the caps and typically the other threads don't have a GL context bound.

As a result we've decided to make GlfContextCaps an explicit step that applications should do.  Typically just after the application calls GlfGlewInit().

GlfContextCaps is supposed to return reasonable defaults, if an application fails to initialize the caps.  During bringup of this change, it was noticed some tests were
crashing because of these defaults.  Therefore, they have been adjusted.

Public Changes:
- Split GlfContextCaps::GetInstance() into a read-only GlfContextCaps::GetInstance() and initialization function GlfContextCaps::InitInstance()
- Generally added GlfContextCaps::InitInstance() to applications and tests.
- Added "const" for fetches of the context caps where needed.
- Some tests were not initializing the shared GL context, which some systems depend on for resource management.
- Some tests were not initializing a GL context at all, but was using the GL subsystem.
- Seperated out GL subsystem initialization from UsdImagingGLEngine::IsHydraEnabled() into its own entity.  GL subsystem initialization now happens the first time a UsdImagingGLEngine is constructed.
- Removed Hdx tests from apple (non-X11) build.  These tests depend on the test GL context, which is implemented for X11 only.  These tests where never run in the cmake project as they require image diff to verify.

(Internal change: 1946474)