Auto detect translucency MaterialTag when opacity input attribute is authored or connected.
This allows us to not have to make two versions of shaders just to hardcode the 'translucent' materialTag in one of them.
In a lot of cases we can determine the translucent tag correctly by looking at a few commonly used inputs that control translucency.
(e.g. UsdPreviewSurface glslfx).

In the future we can consider expanding this by putting the 'translucency' input attribute names in the meta-data of the glslfx.
That makes the material tagging system more flexible.

(Internal change: 1946515)