[Hdx] Migrate HdxColorizeTask to 3-phase task system.
While a previous check-in added the Prepare step to the task, this change migrates the code over to using the 3 phases.

HdxColorizeTask was one of the simpler tasks to migrate.  The main gotcha was the tasks validation code only executed when the task params
changed, but also need the bprim from the render index.  To handle this, the validation state flag was added as a class member.

Changes:
- Added _needsValidation as a member variable.
- Moved render buffer fetch and validation from Sync to the Prepare phase.

(Internal change: 1946751)