[Hdx] Migrate HdxSelectionTask to 3-phase task system.
While a previous check-in added the Prepare step to the task, this change migrates the code over to using the 3 phases.

Pretty much all of the tasks functionality moved to the Prepare step.  The only thing remaining is the fetch of the task params
into the class member.

The change allowed some cleanup of some old behavior of selection task.  For example, it was obtaining the task dirty bits from
the change tracker, rather than using the passed dirtyBits parameter.

HdxSelectionTask also has some change tracking, so it only creates buffers and updates uniforms if the selection state or params
changed.  As the prepare step doesn't have access to the dirty bits (as they are already clean), the code resets the internal selection
version number, in order to force the uniform update.

To match the phase of processing, the Selection Trackers Sync function was renamed Prepare.

Changes:
- HdxSelectionTask no longer gets dirty bits direct from the change tracker.
- DirtyParams dirty bit, now resets the tasks copy of the last selection tracker version
- Migrated code from Sync to Prepare
- Renamed HdxSelectionTracker::Sync to HdxSelectionTracker::Prepare.

(Internal change: 1946752)