[UsdImaging] Fix for single value pull for computation primvar descriptor
We pull on both "authored" and "computation" primvar descriptors in the backends, The former always has an entry, while the latter was populated only by adapters inserting/injecting computations on a prim.
With the recently added support for Hydra computations in UsdImaging, if the value cache didn't have an entry, we delegated to the registered adapter with the DirtyPrimvar bit via _UpdateSingleValue. DirtyPrimvar is an overloaded bit, and triggers a lot more processing than was intended.

The generic case currently is that a prim doesn't have computed primvars, in which case we choose to not populate an entry into the value cache.

(Internal change: 1946867)