[Hdx] Migrate HdxRenderTask and HdxRenderSetupTask to 3-phase task system.
While a previous check-in added the Prepare step to the task, this change migrates the code over to using the 3 phases.

In the setup task, the camera and aov binding sync has moved over to the prepare phase.  As they are part of prepare, they don't need to be explicitly synced by the render pass anymore.

The setups task deferend creation of _renderPassState can now happen in the prepare phase, if SyncParams is never called.

Changes:
- Renamed SyncAovBindings to PrepareAovBindings and moved call to Prepare().
- Renamed SyncCamera to PrepareCamera and moved call to Prepare().
- Remove verify that _renderPassState has been created and switched to _GetRenderPassState() instead.
- _GetRenderPassState(), PrepareAovBindings() and  PrepareCamera now take the render index as a parameter instead of the scene delegate.
- Render Task no longer has explicit calls to SyncAovBindings() and SyncCamera().

(Internal change: 1947160)