Fixes a crash/hang bug with alembic faceset support
In the alembic faceset writer, we were taking a reference to a
temporary SdfAbstractSpecId, which resulted in corrupt memory.
On some platforms that can either manifest as a crash or as a hang
as the path node data structures form circular references and hang
in inner loops.  This was caught by testUsdAbcFaceset on macOS.

(Internal change: 1947657)