[Hdx] Fix for accessing draw target before it is set.
Some internal tasks regressed (asserted) with the 3-phase task changes, because they were using HdxDrawTargetRenderPass and not setting the draw target until the Prepare Phase.

In the future it is hopped that task sync can be parrellel, so with the 3-phase task it was incorrect for HdxDrawTargetRenderPass to assume there is a valid GL context during Sync.
Therefore, this change moves the HdxDrawTargetRenderPass context check to the prepare phase.

Changes:
- Added Prepare() phase to HdxDrawTargetRenderPass
- Moved draw target context check to Prepare()
- Fixed some incorrect comments in HdxDrawTargetRenderPass's header.
- Added calls to HdxDrawTargetRenderPass's Prepare() in HdxDrawTargetTask's Prepare()

(Internal change: 1947778)