[usdMaya/pxrUsdTranslators] Improve error message/clean up logspam when exporting joints that are scene roots
Joints need to be parented under another group in order to be exported;
this was already enforced by the exporter. However, some intermediary
computations with path handling caused a lot of warnings to be printed
out to the Maya console.
This change cleans up the error cases so that the exporter only errors
once inside UsdMaya_SkelBindingsProcessor. I've updated the error text
to be more informative and to tell the user the most likely fix for the
error. In addition, I changed the errors from TF_WARN to TF_RUNTIME_ERROR
since they would already cause the export process to halt; this just has the
effect of changing the yellow status line into a red status line as well as
populating the description of Maya's RuntimeError when invoked from Python.

(Internal change: 1947838)