[Glf] Don't print GL debug messages if they are just reporting debug group push/pop.
Glf has a mechanism to forward driver debug messages to TF_WARN.  At some point, it started forwarding the debug group push/pop messages; the result is that turning on the GL diagnostic trace (which gives really helpful markers in apitrace) also generates an epic amount of spam.

The solution is to ask the driver to stop delivery of that category of message.

(Internal change: 1948431)