pxrTf: Make sure to use correct ordering of base types when declaring TfTypes.
Previously, the order of bases was defined on the go as types were (repeatedly)
declared: A common scenario was that Plug would declare a type using a subset
of the bases and then C++ code would supply all bases.   However,
only additional bases would be added to a TfType's list of bases.  This resulted
in a mangled order of bases.

Many behaviors in the system are based on TfType's ordering of bases, mostly
via TfType::GetAllAncestorTypes().

What made things worse is that the order in which TfType's are declared is
not stable: Sometimes Plug comes first, sometimes the types declared in the
C++ code.

This change makes it so that the ordering will be stable and adds more testing
and error messages if clients attempt to change the base order.

(Internal change: 1948578)