Address performance regression due to purpose cache
When we added the purpose inherited cache in change 1948635, some
of our tests regressed because it cost more to shut down Hydra now
(more cost in destroying the new cache). To address this, we now destroy
the expensive parts of these caches asynchronously.  We now shut down
the USD imaging scene delegate faster than before the regression.
(On the test in question average shut down time went from ~70ms to ~50ms,
before the new cache it was at ~65ms)

(Internal change: 1950029)