Change the FreeCamera auto-clip behavior into an optional mode.
Automatic computation of near/far clipping planes to maximize
use of Z-buffer resolution for a given view frustum is a
heuristic developed for, and maximally benefitting, the
exploration of very large environments.  We've tweaked the
algorithm a couple of times to make it work better in other
scenarios, but given the fundamental restriction that we
cannot recompute the closest-geometry-to-camera every time
we draw (since it could halve the frame rate), it still has
problems with scenes whose extent changes over time,
and for close-up work.  We finally conclude that the behavior
needs to be optional.

We therefore introduce a viewSetting autoComputeClippingPlanes
to track and save the mode, reprupose the "View > Recompute Clipping Planes"
menu item to "View > Auto Compute Clipping Planes" to toggle the mode,
and minimally thread awareness of the mode through StageView
and FreeCamera.

The "Camera:" HUD info will now display "Free AutoClip" when the
mode is active for easy tracking, and for new usdview sessions, the
default is OFF.

We also enabled ToolTips for the View menu and added a tip for
the new command.  This seems like a better and more maintainable
way of documenting (most) usdview functionality, and we hope to
add good tips for other commands in the future.

(Internal change: 1950704)