Clean up Sdf spec type registration
The macros for registering Sdf spec types would add
'hidden' public static API that were only intended
to be used when adding the various spec types to the
registry. Leaking implementation details into public
API like this is not ideal for a number of reasons.

This change reworks the registration mechanism so
these details are hidden away in the .cpp file and
not exposed publicly.

(Internal change: 1955797)