[Hd] Remove Include / Exclude filter from dirty lists.
While implementing the 3-phase task re-factor, an issue was encountered we've dubbed the "Collection Paradox".

What this problem is, there is a race-condition during Hydra between tasks and sprim sync.
The problem is that the current Sync algorithm needed Collection information from tasks in order to determine what to sync.
For some tasks, namely shadows and draw targets, this collection information is obtained from Sprims, which have not been synced yet.
Hence needing to the sync the sprims to know which prims (including sprims) to sync - the paradox.

The issue will be address over a series of commits.

To resolve the paradox, the different prim filters managed by collections is being broken down, such as each type of filter is managed
distinctly by the sync algorithm.

This first commits addresses the path-based filtering (include/exclude).
While the path-based filtering is an important tool for per-render pass operations, sync uses the union of all the filters across all render passes.

After examining all the scene delegates we have access to, it was concluded that in the vast majority of cases, this union represents the complete
set of prims.  Also, typically these paths are authored from the users scene graph perspective. This also presents some challenges for Path-based
filtering.   What happens when we reference a prim outside the filter, such as instancing in Usd?
Some front-ends use aliasing by duplicating prims under different paths which also causes questions of how to interpret these paths.

Therefore, it was decided for the purposes of Sync to disable the path-based filtering.  While recognizing, in the future we may want to consider
alternative data paths and methods for doing scene-level path-based filtering.

Changes:
- Dirty list now uses a hard-coded set of include/exclude paths, rather than obtaining them from the collection.

(Internal change: 1955837)