[Hdx] HdxColorizeSelectionTask for backend-independent selection highlighting.
The task takes as input an HdxSelectionTracker and ID buffers corresponding to the "primId", "elementId", and "instanceId" AOVs.  The task will then colorize the ID buffers with the selection color and blend that onto the current framebuffer.  The intent is to run the main render first, so that this task blends the selection over the normal color output.

(Internal change: 1956196)
(Internal change: 1956488)