[Hd] Remove GetSize() API from HdDirtyList.
This is a small API cleanup.

The HdDirtyList::GetSize() appears on the surface to be a cheap API that would not have side effects.  When in reality it forwards the call to the expensive core GetDirtyRprims() function, which can rebuild the dirty list.

As a result, we would want to really control when GetDirtyRprims() is called and not accidentally trigger the processing.

Fortunately this API was only used by some select white-box tests.  Therefore, I've removed the API to prevent accidental out-of-spec usage.

The GetDirtyRprims() API is going to be changing in a later change for the "collection paradox".  So this also avoid also needing to propagate the change to this API aswell.

Changes:
- Removed HdDirtyList::GetSize().
- Updated tests to call GetDirtyRprims() and get the size for the returned vector.

(Internal change: 1956364)