[usdMaya] refactor util functions for converting Maya node names to SdfPaths and stripping namespaces
This change rewrites UsdMayaUtil::stripNamespaces() to operate on
std::strings instead of MStrings, and makes it preserve the absolute-ness
or relative-ness of the passed in node name. Previously, this function
would always prepend a pipe character, causing all returned node names to
be absolute.

This also adds a UsdMayaUtil::MayaNodeNameToSdfPath() function that takes
on the bulk of the string manipulation work previously done by
MDagPathToUsdPath(). This allows the same path manipulation (minus any
transform and shape merging) to be done from a std::string as can be done
from an MDagPath. This comes in handy when working with assembly edits,
since they will be represented as strings, and the nodes that they refer to
may or may not exist in the scene depending on assembly load states and
currently active representations.

Constants for Maya's DAG node hierarchy and namespace delimiters were added
and are used in the string manipulation that happens in the modified
functions.

(Internal change: 1956437)