[usdMaya] fix for nested assembly contents disappearing in Expanded or Full representations with instancing
When importing on behalf of a scene assembly, we want to make sure that we
traverse down into instances. The expectation is that the user switched an
assembly to its Expanded or Full representation because they want to see
the hierarchy underneath it, possibly with the intention of making
modifications down there. As a result, we don't really want to try to
preserve instancing, but we do need to be able to access the scene
description below the root prim.

Prior to this change, USD would hide all scene description below the
instanceable prim, so the import triggered by the assembly representation
change would see no prims underneath the root prim and wouldn't import
anything, leaving the assembly empty underneath.

(Internal change: 1956440)