[usdMaya] more robust handling of assembly edits and fixes for edits on nested assemblies
This change fixes an issue where assembly edits for nested assemblies would
not get picked up in UsdMayaEditUtil and would therefore not appear to have
any effect on assemblies until they were switched to the "Full"
representation.

Previously, UsdMayaEditUtil::GetEditsForAssembly() was finding the edits
for an assembly by constructing an MItEdits iterator using the assembly as
both the editsOwner and the targetNode. This results in an iterator that
yields edits only under that particular assembly, but not under any nested
subassemblies. The Maya docs suggest that using MObject::kNullObj as the
targetNode should yield all edits stored on the assembly, but that actually
seems to yield all edits in the whole scene, including those *not* intended
for editsOwner. As a result, we get this complete list and then filter it
by using a namespace match of the edited node against the assembly node
we're currently considering.

This change also includes a bunch of miscellaneous cleanup, including
renaming all "RefEdit..." types to "AssemblyEdit..." instead to clarify
their purpose.

(Internal change: 1956630)