[Hd] Remove constness from task list in Hydra API.
This is a small API cleanup.

The standard STL container methods forwards the constness of the container onto the constness of the elements of that container.
The TF_FOR_ALL macro allowed us to break the and gave non-const access to the members of the container.

This mechanism was used to access the tasks as part of the task list, so broke when I tried to access to the tasks via methods
other than TF_FOR_ALL.

Therefore, this change removes the constness from the "tasks" parameter to SyncAll, so that the standard container methods can be used.
As a result, I've removed the TF_FOR_ALL and replaced with a standard for loop, to avoid this issue again.
As this change is already modifying the SyncAll and Execute API, we took the opertunity to fix the code style violation, so that renderIndex is
now passed by pointer rather than reference (as it's non-const).

Changes:
- Removed the const from the tasks param of HdRenderIndex::SyncAll.
- Change tasks and renderIndex to be passed by pointer instead of reference.
- This change needed propagating up the call stack.
- Replaced the task TF_FOR_ALL with a standard for loop.

(Internal change: 1956904)