Introduce the concept on ImageShaderRenderPass in HdSt.
This can be used to render fullscreen post-effects.

ImageShaderRenderPass creates a single drawItem with 3 points that are pushed through the immediate draw batch.
The shader used to render is the RenderPassShader provided by the client task that wishes to render a fullscreen effect.

By using this approach (immediateDrawBatch + RenderPassShader) the client has full control over the shader used to render the effect, but the shader also goes through code-gen.
This means fullscreen effects can access SSBO and uniform buffers registered via the RenderPassShader.

(This is needed for a follow up changelist that introduces a 2-pass order independent transparency)

(Internal change: 1957051)