[usdMaya/pxrUsdTranslators] add import/export support for RfM cylinder lights
This change also re-works the Renderman for Maya light import and export
unit tests to be based on RfM 22.4. In this version, RfM writes all light
and shader attributes into the Maya scene file, so we no longer check that
'shadow:enabled' is unauthored and instead just check that it is true.

(Internal change: 1957318)