[Hd] Move responsibility for dirtying collection on repr changes into the change tracker.
Prior to this change, Hydra depended on the scene delegates to call MarkAllCollectionsDirty if Repr changed.
As the resolved repr is a combination the view and task opinion of repr (in the collection) and repr from the prim,
changing a prim's opinion could change it's resolved repr. Therefore, we need to possibly init the new resolved repr in
Sync and rebuild all the batches.

MarkAllCollectionsDirty is a big hammer way to cause this, but Hydra depended on the Scene delegates to know to do this.
This change moves this logic to within the change tracker. So Scene delegate just need to mark the prims Repr dirty.

Testing of the change revealed an existing bug in MarkRprimDirty.  When a prim is filtered out from the Sync, it leaves the
dirty bits for the prim set, so changes can be pulled the next time the prim is included.
However, if the prim got filtered out and then a prim changed a filter parameter (Repr or Render Tag), MarkRprimDirty()
was early exiting.  If it was the only change in the scene, the dirty list would think no changes had be made at all and happily skip
any sync processing.

Therefore, it is not appropriate to do the early-out if nothing changes for one of the filter params.
Repr and Render Tag requires render index version to be updated to cause collections to be rebuilt
and change count need sto be updated (to avoid the early-out in dirty list).

Changes:
- If DirtyRepr is set in MarkRprimDirty increment _indexVersion, which will increase all the collections versions.
- If Visibility, Repr or Render Tag is dirty.  Skip the early out in MarkRprimDirty.
- Updated MarkAllRprimDirty to match MarkRprimDirty behaviour

(Internal change: 1957591)